Fix screenshot handeling in WTF.
Initial assumption was that the webworker could deal with images. Of course it can't, so in order to handle images we need to transfer the URL to the mainthread.
Have the mainthread fetch the image, stick it in a canvas object and return the imageData back to the worker.

This checkin adds the above flow. Did not test for leaks, so beware in running endurance tests.
However for now this unblocks all the screenshot-savvy tests.

Signed-off-by: wouterlucas <wouter@wouterlucas.com>